+++ src/endoscope/option/options_reconstructor.hpp
@@ -59,7 +59,7 @@ bool parse_command_options(
     bool *est_move = nullptr,
     float *thresh_knn_ratio = nullptr, float *thresh_ransac = nullptr,
     int *cpu_core = nullptr, size_t *scene = nullptr,
-    size_t *matching = nullptr, int *extractor = nullptr,
+    size_t *matching = nullptr, size_t *extractor = nullptr,
     size_t *publish = nullptr);
 
 #endif // RECONSTRUCTION__OPTIONS_HPP_
